+++ darts/tests/models/forecasting/test_local_forecasting_models.py
@@ -8,15 +8,15 @@ from unittest.mock import Mock, patch
 import numpy as np
 import pandas as pd
 import pytest
+from statsforecast.models import AutoARIMA as SFAutoARIMA
 
 from darts.datasets import AirPassengersDataset, IceCreamHeaterDataset
 from darts.logging import get_logger
 from darts.metrics import mape
 from darts.models import (
     ARIMA,
-    # BATS,
     FFT,
-    # TBATS,
+    TBATS,
     VARIMA,
     AutoARIMA,
     AutoCES,
@@ -36,6 +36,7 @@ from darts.models import (
     Prophet,
     RandomForest,
     RegressionModel,
+    StatsForecastModel,
     Theta,
 )
 from darts.models.forecasting.forecasting_model import (
@@ -60,11 +61,12 @@ models = [
     (ARIMA(12, 2, 1), 5.2),
     (ARIMA(1, 1, 1), 24),
     (AutoARIMA(season_length=12), 4.6),
+    (StatsForecastModel(SFAutoARIMA(season_length=12)), 4.6),
     (AutoTheta(season_length=12), 5.5),
     (AutoCES(season_length=12, model="Z"), 7.3),
     (AutoETS(season_length=12, model="AAZ"), 7.3),
     (AutoMFLES(season_length=12, test_size=12), 9.8),
-    (AutoTBATS(season_length=12), 10),
+    (AutoTBATS(season_length=12), 10.0),
     (Croston(version="classic"), 23),
     (Croston(version="tsb", alpha_d=0.1, alpha_p=0.1), 23),
     (Theta(), 11),
@@ -79,8 +81,10 @@ models = [
     (KalmanForecaster(dim_x=3), 20),
     (LinearRegressionModel(lags=12), 13),
     (RandomForest(lags=12, n_estimators=5, max_depth=3), 14),
-    # (TBATS(use_trend=True, use_arma_errors=True, use_box_cox=True), 8.5),
-    # (BATS(use_trend=True, use_arma_errors=True, use_box_cox=True), 11),
+    (
+        TBATS(season_length=12, use_trend=True, use_arma_errors=True, use_boxcox=True),
+        10,
+    ),
 ]
 
 # forecasting models with exogenous variables support
@@ -143,7 +147,12 @@ class TestLocalForecastingModels:
     def test_save_model_parameters(self):
         # model creation parameters were saved before. check if re-created model has same params as original
         for model, _ in models:
-            assert model._model_params == model.untrained_model()._model_params
+            # take string values since StatsForecastModel has `model` as input which does not have `__eq__`
+            model_orig_params = {k: str(v) for k, v in model._model_params.items()}
+            model_fresh_params = {
+                k: str(v) for k, v in model.untrained_model()._model_params.items()
+            }
+            assert model_fresh_params == model_orig_params
 
     @pytest.mark.parametrize("model", [ARIMA(1, 1, 1)])
     def test_save_load_model(self, tmpdir_module, model):
@@ -627,12 +636,15 @@ class TestLocalForecastingModels:
                 ),  # data structure param
                 "KalmanForecaster(add_encoders={'cyclic': {'past': ['month']}})",
             ),
-            # (
-            #     TBATS(
-            #         use_trend=True, use_arma_errors=True, use_box_cox=True
-            #     ),  # params in wrong order
-            #     "TBATS(use_box_cox=True, use_trend=True)",
-            # ),
+            (
+                TBATS(
+                    season_length=12,
+                    use_trend=True,
+                    use_arma_errors=True,
+                    use_boxcox=True,
+                ),
+                "TBATS(season_length=12, use_trend=True, use_arma_errors=True, use_boxcox=True)",
+            ),
         ],
     )
     def test_model_str_call(self, config):